remove internal component
