+++ src/components/AutoComplete/AutoComplete.tsx
@@ -1,8 +1,13 @@
+import { useCombobox } from "downshift";
 import { nanoid } from "nanoid";
 import React, { useMemo } from "react";
 import { useMergedProps } from "../../hooks/useMergedProps";
+import useTheme from "../../hooks/useTheme";
+import Icon from "../Icon";
 import Field from "../internal/Field";
-import InternalAutoComplete from "../internal/InternalAutoComplete";
+import InternalInput from "../internal/InternalInput";
+import LoadingIcon from "../LoadingIcon";
+import VisuallyHidden from "../VisuallyHidden";
 import { defaultAutoCompleteProps } from "./defaultAutoCompleteProps";
 import { AutoCompleteProps, ListItemKey } from "./types";
 import useGetItems from "./useGetItems";
@@ -28,6 +33,7 @@ function AutoComplete<Item extends ListItemKey = ListItemKey>(
   props: AutoCompleteProps<Item>
 ) {
   const mergedProps = useMergedProps(props, defaultAutoCompleteProps);
+  const theme = useTheme();
 
   const {
     label,
@@ -38,10 +44,10 @@ function AutoComplete<Item extends ListItemKey = ListItemKey>(
     onFocus,
     disabled,
     helpText,
-    itemToString,
+    itemToString: itemToStringFn,
     placeholder,
-    itemsFooter,
-    listItem,
+    itemsFooter: Footer,
+    listItem: ListItem,
     hideLabel,
     testId,
     value,
@@ -56,7 +62,51 @@ function AutoComplete<Item extends ListItemKey = ListItemKey>(
 
   const auxId = useMemo(() => `auto-complete-aux-${nanoid()}`, []);
   const { fieldErrors, hasErrors } = getFieldErrors(error);
-  const describeBy = helpText || hasErrors ? auxId : undefined;
+  const describedBy = helpText || hasErrors ? auxId : undefined;
+
+  const itemToString = (item: Item | null): string =>
+    itemToStringFn ? itemToStringFn?.(item) : item ? String(item) : "";
+
+  const {
+    isOpen,
+    getMenuProps,
+    getInputProps,
+    getItemProps,
+    getComboboxProps,
+    getToggleButtonProps,
+    highlightedIndex,
+    inputValue,
+    selectItem,
+    closeMenu,
+  } = useCombobox<Item | null>({
+    items: items || [],
+    defaultSelectedItem: value,
+    onInputValueChange: (changed) => {
+      getItems(changed);
+    },
+    onSelectedItemChange: (changed) => {
+      onChange?.(changed.selectedItem);
+    },
+    itemToString: (item) =>
+      itemToStringFn ? itemToStringFn?.(item) : item ? String(item) : "",
+  });
+
+  const onClear = () => {
+    selectItem(null);
+    closeMenu();
+  };
+
+  const menuIsOpen = isOpen || (__internal__open ?? false);
+  const showClearIcon = (menuIsOpen && items.length > 0) || inputValue !== "";
+  const showMagnifier = !showClearIcon;
+  const loading = status === "LOADING" || __internal__loading;
+
+  const renderListItem = (record: Item) => {
+    if (ListItem) {
+      return <ListItem inputValue={inputValue} item={record} />;
+    }
+    return itemToString ? itemToString(record) : record;
+  };
 
   return (
     <Field
@@ -69,26 +119,76 @@ function AutoComplete<Item extends ListItemKey = ListItemKey>(
       errors={fieldErrors}
       testId={testId}
     >
-      <InternalAutoComplete<Item>
-        label={label}
-        innerRef={innerRef}
-        onBlur={onBlur}
-        onFocus={onFocus}
-        items={items}
-        isLoading={status === "LOADING"}
-        itemToString={itemToString}
-        onChange={onChange}
-        placeholder={placeholder}
-        onInputValueChange={getItems}
-        value={value}
-        itemsFooter={itemsFooter}
-        listItem={listItem}
-        describedBy={describeBy}
-        __internal__open={__internal__open}
-        __internal__highlightedIndex={__internal__highlightedIndex}
-        __internal__loading={__internal__loading}
-        __internal__focus={__internal__focus}
-      />
+      <div css={theme.autoComplete.getCSS({ targetElement: "container" })}>
+        <div {...getComboboxProps()}>
+          <InternalInput
+            {...getInputProps({ refKey: "innerRef", ref: innerRef })}
+            label={label}
+            describedBy={describedBy}
+            onFocus={onFocus}
+            onBlur={onBlur}
+            placeholder={placeholder}
+            __internal__focus={__internal__focus}
+            autoComplete="off"
+            hasSuffixButton
+            isLoading={loading}
+          />
+        </div>
+        <div css={theme.autoComplete.getCSS({ targetElement: "right" })}>
+          {loading && <LoadingIcon />}
+
+          <button
+            type="button"
+            onClick={onClear}
+            tabIndex={-1}
+            css={theme.autoComplete.getCSS({
+              targetElement: "clearIcon",
+              showClearIcon: showClearIcon,
+            })}
+          >
+            <Icon name="cross" />
+            <VisuallyHidden>Clear</VisuallyHidden>
+          </button>
+
+          {showMagnifier && (
+            <button
+              type="button"
+              {...getToggleButtonProps()}
+              css={theme.autoComplete.getCSS({ targetElement: "searchIcon" })}
+            >
+              <Icon name="search" />
+              <VisuallyHidden>Search</VisuallyHidden>
+            </button>
+          )}
+        </div>
+        <ul
+          {...getMenuProps()}
+          css={theme.autoComplete.getCSS({
+            targetElement: "ul",
+            isOpen: menuIsOpen,
+          })}
+        >
+          {menuIsOpen && (
+            <>
+              {items.map((record, index) => (
+                <li
+                  key={record.id || index}
+                  css={theme.autoComplete.getCSS({
+                    targetElement: "li",
+                    isHighlighted:
+                      highlightedIndex === index ||
+                      __internal__highlightedIndex === index,
+                  })}
+                  {...getItemProps({ index, item: record })}
+                >
+                  {renderListItem(record)}
+                </li>
+              ))}
+              {Footer && <Footer closeMenu={closeMenu} />}
+            </>
+          )}
+        </ul>
+      </div>
     </Field>
   );
 }
